Library of generic Files
Because some files might be useful for multiple projects I decided to collect them here and make them more maintainable.
In case they need to be connected functionally to the project, there are configuration functions, that can be called from the project.
For example:
setGetBuffListFunction(GetBuffListFunction getBuffList)
Allows you to configure a function that returns the bufflist for a given unit.