Major Bug Explanation + variable name change
Before the commit "Buff Update" (aad759de12a5e7734e24e5ce9e77691a1f70c176)
an iterator was manually used instead of a for element e in list loop.
Because of the return inside the loop, the iterator was potentially not closed, resulting in a leak.
If using the for loop instead, wurst is smart enough to close the iterator before the return statement.